+++ core/src/main/scala/org/opentorah/xml/PrettyPrinter.scala
@@ -40,6 +40,7 @@ final class PrettyPrinter(
     preformattedElements
   )
 
+  @scala.annotation.targetName("add")
   def +(other: PrettyPrinter): PrettyPrinter = copy(
     doNotStackElements = doNotStackElements ++ other.doNotStackElements,
     alwaysStackElements = alwaysStackElements ++ other.alwaysStackElements,